% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_func.R
\name{leaf_background}
\alias{leaf_background}
\title{Leaflet background}
\usage{
leaf_background(x, token = NULL)
}
\arguments{
\item{x}{leaflet object}

\item{token}{Lantmäteriverket user token, otherwise uses option "leaf.token"}
}
\description{
Add leaflet background tiles from lantmäteriverket topotgraphical map in EPSG:3006 projectopmn
}
