% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_func.R
\name{leaf_marker}
\alias{leaf_marker}
\title{Leaflet marker wrapper}
\usage{
leaf_marker(
  x,
  d,
  lbl = NULL,
  interactive = FALSE,
  color = NULL,
  size = NULL,
  colorby = NULL,
  opacity = 1
)
}
\arguments{
\item{x}{leaflet object}

\item{d}{polygon data, sf object}

\item{lbl}{name of variable to use as label}

\item{interactive}{boolean}

\item{size}{size of circle}

\item{colorby}{name of variable to use for fill color}

\item{opacity}{between 0 and 1}
}
\description{
Add marker to leaflet
}
