% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_func.R
\name{leaf_polygon}
\alias{leaf_polygon}
\title{Leaflet polygon wrapper}
\usage{
leaf_polygon(
  x,
  d,
  lbl = NULL,
  interactive = FALSE,
  colorby = NULL,
  opacity = 1
)
}
\arguments{
\item{x}{leaflet object}

\item{d}{polygon data, sf object}

\item{lbl}{name of variable to use as label}

\item{interactive}{boolean}

\item{colorby}{name of variable used to fill polygons}

\item{opacity}{between 0 and 1}
}
\description{
Add polygon to leaflet
}
